requirements.txt
streamlit
langchain-google-genai
langchain-core
PyMuPDF
python-docx
pandas
fpdf=
PyYAML
